requirements.txt
streamlit
pandas
numpy
plotly
diffusers
transformers
accelerate
torch
groq
python-dotenv

